cadstrando contas

+++ src/model/Lancamento.ts
@@ -1,7 +1,7 @@
 
 import { IDefault, Inject } from './IDefault';
 import * as mongoose from 'mongoose';
-import { Double, Int32 } from 'bson';
+import { Double, Int32, Decimal128 } from 'bson';
 
 export interface ILancamentoModel extends IDefault, mongoose.Document{
     id: string;
@@ -16,10 +16,10 @@ export interface ILancamentoModel extends IDefault, mongoose.Document{
 let schema = {
     pagador: {type: mongoose.Schema.Types.ObjectId, ref: 'contacorrente', required: true},
     recebedor: {type: mongoose.Schema.Types.ObjectId, ref: 'contacorrente', required: true},
-    valorTotal: {type: Double, required: true},
+    valorTotal: {type: Decimal128, required: true},
     parcelas: {type: Int32, required: true},
-    saldoPagador: {type: Double, required: true},
-    saldoRecebedor: {type: Double, required: true}
+    saldoPagador: {type: Decimal128, required: true},
+    saldoRecebedor: {type: Decimal128, required: true}
 };
 
 Inject(schema);
